Broke out baseline correction implementation into separate file, preparing to later select between different methods. Broke out peak_fitting_helpers.
