+++ prisma/__main__.py
@@ -5,10 +5,11 @@ import prisma
 
 
 def main():
-    
-    prisma_path = '/'.join(prisma.__path__[0].split('\\')[:-1]) #fin path where prisma is installed
-    gui_path = prisma_path +'/gui/'
-    subprocess.run(['voila', gui_path + 'GUI.ipynb']) #Open GUI with voila
+    # Find path where prisma is installed
+    prisma_path = '/'.join(prisma.__path__[0].split('\\')[:-1])
+    gui_path = prisma_path + '/gui/'
+    # Open GUI with voila
+    subprocess.run(['voila', gui_path + 'GUI.ipynb'])
 
 
 if __name__ == '__main__':
