+++ prisma/spectrum.py
@@ -2,11 +2,9 @@
 import numpy as np
 import math
 import scipy as sp
-from scipy import sparse
-from scipy.sparse.linalg import spsolve
 from typing import Optional
 from scipy.optimize import curve_fit
-import prisma.util.lineshapes
+import prisma.util.peak_fitting_helpers
 
 """Templates of Spectrum object"""
 
@@ -149,8 +147,8 @@ class Spectrum:
         else:
             return spectrum
 
-    def asymmetric_least_squares(self, log_p=-1.5, log_lambda=7):
-        PROCESS_TYPE = "Asymmetric Least Squares"
+    def baseline_correction(self, log_p=-1.5, log_lambda=7):
+        PROCESS_TYPE = "Baseline Correction"
         BATTINFO_ID = self.object_identifiers["BattInfo ID"]
         new_metadata = {
             "Process": PROCESS_TYPE,
@@ -160,45 +158,13 @@ class Spectrum:
             "Log10(lambda)": log_lambda,
         }
         spectrum = self
-        param_p, param_lambda = 10**log_p, 10**log_lambda
-
-        nan_counts_mask = np.isnan(spectrum.counts)
-        if np.any(nan_counts_mask):  # interpolation of nans
-
-            non_nan_counts = np.interp(
-                np.arange(len(spectrum.counts)),
-                np.arange(len(spectrum.counts))[~nan_counts_mask],
-                spectrum.counts[~nan_counts_mask],
+        baseline_correction = (
+            prisma.util.baseline_corrections.asymmetric_least_squares(
+                spectrum.counts, log_p, log_lambda
             )
-        else:
-            non_nan_counts = spectrum.counts
-
-        m = len(non_nan_counts)
-        D = sparse.diags(
-            [1, -2, 1], [0, -1, -2], shape=(m, m - 2)
-        )  # sparse representation of difference_2 matrix
-        w = np.ones(m)  # initial -symmetric- weights
-        W = sparse.spdiags(w, 0, m, m)  # weight matrix with initial weights
-        iterations = 20
-
-        for _ in range(iterations):
-
-            W.setdiag(w)  # wiegth matrix is updated with newest weights
-            C = W + param_lambda * D.dot(
-                D.transpose()
-            )  # matrix summarizing the fit and smooth penalties
-            z = spsolve(C, w * non_nan_counts)
-            updated_w = param_p * (non_nan_counts > z) + (1 - param_p) * (
-                non_nan_counts < z
-            )
-
-            if np.linalg.norm(w) == np.linalg.norm(updated_w):
-                break
-            else:
-                w = updated_w
-
-        new_counts = non_nan_counts - z
-        new_counts[nan_counts_mask] = np.nan
+        )
+        z = baseline_correction[1]
+        new_counts = baseline_correction[0]
         new_indexes = spectrum.indexes
 
         return Spectrum(
@@ -207,84 +173,6 @@ class Spectrum:
             baseline=z,
             metadata=new_metadata
         )
-    # *****************************HELPER FUNCTIONS************************
-
-    def prisma_peak_defaults(self, peak_bounds, max_widths):
-        # Format peak_bounds and peak_widhts to parameter bounds for the curve
-        # fit function
-        # init_guess   --> initial guesses for the fitting parameters:
-        # [y0, h1, p1, w1, h2, p2, w2, h3, p3, w3, ...]
-        # initial guesses for the fitting parameters
-        # param_bounds --> 2-tuple of lists with lower and upper bounds for
-        # the fitting parameters: ([y0,h1,p1,w1,...],[y0,h1,p1,w1,...])
-        spectrum = self
-
-        overall_max_counts = np.amax(spectrum.counts)
-
-        limit_resolvable_width = self.RESOLVABLE_WIDTH_FACTOR * np.abs(
-            spectrum.indexes[1] - spectrum.indexes[0]
-        )
-
-        # Bounds for y0
-        init_guess = [0]
-        param_bounds_low = [-0.1 * overall_max_counts]
-        param_bounds_high = [0.1 * overall_max_counts]
-
-        # Bounds for all other parameters
-        for width, bound in zip(max_widths, peak_bounds):
-
-            max_counts_within_bounds = np.amax(
-                spectrum.counts[
-                    (spectrum.indexes > bound[0]) &
-                    (spectrum.indexes < bound[1])
-                ]
-            )
-
-            # guess height = 30% maximum height
-            # guess position: halfway between bounds
-            # guess width: half the maximum width provided or 5% more of
-            # min_resolvable_width, whoever is greater
-            init_guess += [
-                0.3 * max_counts_within_bounds,
-                0.5 * (bound[1] - bound[0]) + bound[0],
-                max(1.05 * limit_resolvable_width, width / 2),
-            ]
-
-            # lower bound height = 0 | lower bound position: the one provided |
-            # lower bound width: minimum resolvable width
-            param_bounds_low += [0, bound[0], limit_resolvable_width]
-
-            # upper bound height = 110% max height | upper bound position:
-            # the one provided | upper bound width: the one provided or 10%
-            # more of min_resolvable_width, whoever is greater
-            param_bounds_high += [
-                1.1 * max_counts_within_bounds,
-                bound[1],
-                max(1.1 * limit_resolvable_width, width),
-            ]
-
-        return init_guess, (param_bounds_low, param_bounds_high)
-
-    def get_fitting_functions(self, lineshape, number_of_peaks):
-        if lineshape == 'Lorentzian':
-            fitting_function = prisma.util.lineshapes.lorentzians(
-                number_of_peaks
-            )
-            single_peak_function = prisma.util.lineshapes.lorentzians(1)
-
-        elif lineshape == 'Gaussian':
-            fitting_function = prisma.util.lineshapes.gaussians(
-                number_of_peaks
-            )
-            single_peak_function = prisma.util.lineshapes.gaussians(1)
-
-        elif lineshape == 'Pseudo-Voight 50% Lorentzian':
-            fitting_function = prisma.util.lineshapes.pseudo_voight_50(
-                number_of_peaks
-            )
-            single_peak_function = prisma.util.lineshapes.pseudo_voight_50(1)
-
-        return fitting_function, single_peak_function
 
     # ***************************FITTING FUNCTION***************************
 
@@ -310,12 +198,15 @@ class Spectrum:
         new_indexes = spectrum.indexes
 
         # formatting bounds and define fitting functions with helper functions
-        init_guess, param_bounds = self.prisma_peak_defaults(
-            peak_bounds, guess_widths
+        init_guess, param_bounds = (
+            prisma.util.peak_fitting_helpers.prisma_peak_defaults(
+                peak_bounds, guess_widths
+            )
         )
-        fitting_function, single_peak_function = self.get_fitting_functions(
-            lineshape_peak,
-            new_metadata["Number of peaks"]
+        fitting_function, single_peak_function = (
+            prisma.util.peak_fitting_helpers.get_fitting_functions(
+                lineshape_peak, new_metadata["Number of peaks"]
+            )
         )
 
         # fitting
